textbook/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
openai
sqlalchemy
psycopg2-binary
pydantic
python-multipart
bcrypt
